Fix data type in IK since we no longer uses motor ticks internally.

+++ robotis_mini/include/robotis_mini/ik.hpp
@@ -5,9 +5,9 @@
 
 namespace robotis_mini::ik {
 
-void IK_RH(float, float, float, std::vector<uint16_t>&);
-void IK_LH(float, float, float, std::vector<uint16_t>&);
-void IK_RF(float, float, float, float roll, float pitch, std::vector<uint16_t>&);
-void IK_LF(float, float, float, float roll, float pitch, std::vector<uint16_t>&);
+void IK_RH(float, float, float, std::vector<float>&);
+void IK_LH(float, float, float, std::vector<float>&);
+void IK_RF(float, float, float, float roll, float pitch, std::vector<float>&);
+void IK_LF(float, float, float, float roll, float pitch, std::vector<float>&);
 
 }
